find optimal compiled mds circuit

+++ twenty-first/src/shared_math/tip5.rs
@@ -491,11 +491,11 @@ impl Tip5 {
     fn mds_generated(state: &mut [BFieldElement; STATE_SIZE]) {
         let mut result = [BFieldElement::zero(); STATE_SIZE];
 
-        let mut lo: [i64; STATE_SIZE] = [0; STATE_SIZE];
-        let mut hi: [i64; STATE_SIZE] = [0; STATE_SIZE];
+        let mut lo: [u64; STATE_SIZE] = [0; STATE_SIZE];
+        let mut hi: [u64; STATE_SIZE] = [0; STATE_SIZE];
         for (i, b) in state.iter().enumerate() {
-            hi[i] = (b.raw_u64() >> 32) as i64;
-            lo[i] = (b.raw_u64() as u32) as i64;
+            hi[i] = (b.raw_u64() >> 32);
+            lo[i] = (b.raw_u64() as u32) as u64;
         }
 
         lo = generated_function(&lo);
